Move event handlers into Slack.Event module
